requirements.txt
selenium==3.141.0
tqdm==4.61.1
Pillow